Removed dependency on Triton Bench
