+++ op_tests/op_benchmarks/triton/bench_moe_mx.py
@@ -9,7 +9,7 @@ from aiter.ops.triton.moe_op_mxfp4 import fused_moe_mxfp4
 from op_benchmarks.triton.utils.common import str_to_torch_dtype, torch_to_tl_dtype
 from op_benchmarks.triton.utils.moe import generate_moe_alignment
 from op_tests.triton_tests.test_moe_mx import alloc_rand
-from triton_kernels.numerics_details.mxfp import downcast_to_mxfp
+from op_tests.triton_tests.utils.quant_ref import torch_dynamic_mxfp4_quant
 
 
 def model_benchmark_configs(args):
@@ -66,7 +66,6 @@ def run_benchmark(args):
         is_a_mixed_input = a_dtype_str.startswith("mx")
         is_b_mixed_input = b_dtype_str.startswith("mx")
         a_dtype = str_to_torch_dtype[a_dtype_str]
-        b_dtype = str_to_torch_dtype[b_dtype_str]
         c_dtype = torch.bfloat16 if is_a_mixed_input else a_dtype
         fp16_dtype = torch.float16 if a_dtype_str == "fp16" else torch.bfloat16
         a_tri = alloc_rand((M, K), dtype=fp16_dtype, device="cuda", requires_grad=False)
@@ -96,18 +95,20 @@ def run_benchmark(args):
         )
         # Downcast a tensor to mxfp4 and upcast back for reference
         if is_a_mixed_input:
-            swizzle_axis = 0 if swizzle_mx else None
-            a_tri, a_mx_scales, _ = downcast_to_mxfp(
-                a_tri, a_dtype, axis=1, swizzle_axis=swizzle_axis
-            )
+            # swizzle_axis = 0 if swizzle_mx else None
+            # a_tri, a_mx_scales, _ = downcast_to_mxfp(
+            #    a_tri, a_dtype, axis=1, swizzle_axis=swizzle_axis
+            # )
+            a_tri, a_mx_scales = torch_dynamic_mxfp4_quant(a_tri)
         else:
             a_mx_scales = None
         # Downcast b tensor to mxfp4 and upcast back for reference
         if is_b_mixed_input:
-            swizzle_axis = 1 if swizzle_mx else None
-            b_tri, b_mx_scales, _ = downcast_to_mxfp(
-                b_tri, b_dtype, axis=2, swizzle_axis=swizzle_axis
-            )
+            # swizzle_axis = 1 if swizzle_mx else None
+            # b_tri, b_mx_scales, _ = downcast_to_mxfp(
+            #    b_tri, b_dtype, axis=2, swizzle_axis=swizzle_axis
+            # )
+            b_tri, b_mx_scales = torch_dynamic_mxfp4_quant(a_tri)
         # (M, K) * (top_k, N, K) -> (M, top_k, N). 2 for multiplication and accumulation
         flops = 2.0 * M * top_k * K * N
         # The weight is applied on the gemm product which has the shape of (M, top_k, N)
